post 1차수정본

+++ src/main/java/com/example/clone_project/service/PostService.java
@@ -2,6 +2,7 @@ package com.example.clone_project.service;
 
 import com.example.clone_project.dto.request.PostRequestDto;
 import com.example.clone_project.dto.response.PostResponseDto;
+import com.example.clone_project.entity.Member;
 import com.example.clone_project.entity.Post;
 import com.example.clone_project.repository.PostRepository;
 import lombok.AllArgsConstructor;
@@ -16,9 +17,11 @@ public class PostService {
 
     private final PostRepository postRepository;
 
-    public Post creatPosts(String content, String file ) {
-        Post post = Post.builder().content(content).imageUrl(file).build();
-        return postRepository.save(post);
+    @Transactional
+    public PostResponseDto creatPosts(Member member , String content, String file ) {
+        Post post = Post.builder().member(member).content(content).imageUrl(file).build();
+        Post postsaved = postRepository.save(post);
+        return new PostResponseDto(postsaved);
 
     }
 
@@ -39,14 +42,20 @@ public class PostService {
 
 
     @Transactional
-    public PostResponseDto updatePosts(Long postId, PostRequestDto postRequestDto) {
+    public PostResponseDto updatePosts(Member member, Long postId, PostRequestDto postRequestDto) {
         Post post = postRepository.findById(postId).orElseThrow();
+        if(!member.equals(post.getMember())) {
+            throw new IllegalArgumentException("사용자 권한이 없습니다.");
+        }
         post.update(postRequestDto);
         return new PostResponseDto(postRepository.save(post));
     }
     @Transactional
-    public void deletePosts(Long postId) {
+    public void deletePosts(Member member, Long postId) {
         Post post = postRepository.findById(postId).orElseThrow();
+        if(!member.equals(post.getMember())){
+            throw new IllegalArgumentException("사용자 권한이 없습니다.");
+        }
         postRepository.delete(post);
     }
 }
